# docs/swagger.yaml
basePath: /api/v1
definitions:
  helpers.HTTPError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  models.Group:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      provider:
        type: string
      updated_at:
        type: string
      users:
        items:
          type: string
        type: array
    required:
    - name
    - provider
    type: object
  models.GroupUser:
    properties:
      id:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
...
  models.ResponseMessage:
    properties:
      message:
        example: runner successfully deleted
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      groups:
        items:
          type: string
        type: array
      id:
        type: string
      password:
        type: string
      provider:
        type: string
      updated_at:
        type: string
      username:
        type: string
    required:
    - password
    - provider
    - username
    type: object
  models.UserGroup:
    properties:
      id:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
...
info:
  contact:
    email: info@kubecode.io
    name: Kubecode Support
    url: https://www.kubecode.io/contact
  description: API Gateway for your kubernetes services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Kriten
  version: v0.3
paths:
...
  /groups/{id}/users:
    delete:
      consumes:
      - application/json
      description: Remove users from group
      parameters:
      - description: Users to be removed
        in: body
        name: group
        required: true
        schema:
          items:
            $ref: '#/definitions/models.GroupUser'
          type: array
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Remove users
      tags:
      - groups
    get:
      consumes:
      - application/json
      description: List all users in given group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.GroupUser'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: List users
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Add users to group
      parameters:
      - description: Users to be added
        in: body
        name: group
        required: true
        schema:
          items:
            $ref: '#/definitions/models.GroupUser'
          type: array
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Add users
      tags:
      - groups
...
  /users:
    get:
      consumes:
      - application/json
      description: List all users available on the cluster
      parameters:
      - description: Maximum number of users to return (default 100)
        in: query
        name: limit
        type: integer
      - description: Number of users to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by user name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Add a user to the cluster
      parameters:
      - description: New user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get groups memberships for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserGroup'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Get user groups
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update a user in the cluster
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Update a user
      tags:
      - users
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"